Fixing dockerfile, adding data to windmill hover, file refactor
