+++ package.json
@@ -25,6 +25,6 @@
         "react": "^18.0.0",
         "react-dom": "^18.0.0",
         "react-router-dom": "^6.3.0",
-        "three": "^0.172.0"
+        "three": "^0.173.0"
     }
 }
